chore(test): extending existing Tab Bar Item Title test screen and scenario with configuration for Android and scenario steps for it
