--- apps/src/tests/single-feature-tests/tabs/test-tabs-item-title-ios/scenario-description.ts
@@ -1,12 +1,0 @@
-import type { ScenarioDescription } from '@apps/tests/shared/helpers';
-
-export const scenarioDescription: ScenarioDescription = {
-  name: 'Tab Bar Item Title',
-  key: 'test-tabs-item-title-ios',
-  details:
-    'Exercises every iOS tab bar item title prop: title, font color,' +
-    ' font family, size, style, weight, and position adjustment.',
-  platforms: ['ios'],
-  e2eCoverage: 'incomplete',
-  smokeTest: false,
-};
